Add a Preferred flag to CompletionItem for IntelliCode
And some clean ups
