+++ src/Features/Core/Portable/Completion/CompletionItemFlags.cs
@@ -26,6 +26,15 @@ namespace Microsoft.CodeAnalysis.Completion
         /// </summary>
         Expanded = 0x2,
 
+        /// <summary>
+        /// Indicates this <see cref="CompletionItem"/> is a preferred item.
+        /// This is determined by the ★ prefix of display text as a temporarily measure to support preference of 
+        /// IntelliCode items comparing to non-IntelliCode items. We expect that Editor will introduce this support
+        /// and we will get rid of relying on the "★" then. We check both the display text and the display text
+        /// prefix to account for IntelliCode item providers that may be using the prefix to include the ★.
+        /// </summary>
+        Preferred = 0x4,
+
         CachedAndExpanded = Cached | Expanded,
     }
 
@@ -36,5 +45,16 @@ namespace Microsoft.CodeAnalysis.Completion
 
         public static bool IsExpanded(this CompletionItemFlags flags)
             => (flags & CompletionItemFlags.Expanded) != 0;
+
+        public static bool IsPreferredItem(this CompletionItem completionItem)
+            => (completionItem.Flags & CompletionItemFlags.Preferred) != 0;
+
+        public static CompletionItem MarkPreferredItem(this CompletionItem item)
+        {
+            if (item.DisplayText.StartsWith("★") || item.DisplayTextPrefix.StartsWith("★"))
+                item.Flags |= CompletionItemFlags.Preferred;
+
+            return item;
+        }
     }
 }
