+++ src/Features/Core/Portable/ExternalAccess/Pythia/Api/PythiaCompletionProviderBase.cs
@@ -29,7 +29,8 @@ namespace Microsoft.CodeAnalysis.ExternalAccess.Pythia.Api
             ImmutableDictionary<string, string>? properties = null,
             ImmutableArray<string> tags = default,
             string? inlineDescription = null)
-            => CommonCompletionItem.Create(displayText, displayTextSuffix, rules, (Glyph?)glyph, description, sortText, filterText, showsWarningIcon, properties, tags, inlineDescription);
+            => CommonCompletionItem.Create(displayText, displayTextSuffix, rules, (Glyph?)glyph, description, sortText, filterText, showsWarningIcon, properties, tags, inlineDescription)
+            .MarkPreferredItem();
 
         public static CompletionItem CreateSymbolCompletionItem(
             string displayText,
@@ -43,7 +44,7 @@ namespace Microsoft.CodeAnalysis.ExternalAccess.Pythia.Api
             ImmutableDictionary<string, string>? properties = null,
             ImmutableArray<string> tags = default)
             => SymbolCompletionItem.CreateWithSymbolId(displayText, displayTextSuffix: null, symbols, rules, contextPosition, sortText, insertionText,
-                filterText, displayTextPrefix: null, inlineDescription: null, glyph: null, supportedPlatforms, properties, tags);
+                filterText, displayTextPrefix: null, inlineDescription: null, glyph: null, supportedPlatforms, properties, tags).MarkPreferredItem();
 
         public static ImmutableArray<SymbolDisplayPart> CreateRecommendedKeywordDisplayParts(string keyword, string toolTip)
             => RecommendedKeyword.CreateDisplayParts(keyword, toolTip);
